backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai>=0.6.0
google-genai>=0.1.0
httpx==0.25.2
pillow==10.1.0
python-multipart==0.0.6
alembic==1.13.0
opencv-python>=4.8.0
ffmpeg-python==0.2.0
openai>=1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
